backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pymongo==4.3.3
python-dotenv==1.0.0
bcrypt==4.0.1
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0
pydantic==1.10.7